Problem: Different instances of a CFML application that are intended to be clustered (using sessionCluster=true) could only be clustered if the different instances have the same application name. two applications could be configured differently but still need to share session storage so they can't have the same application name.
Fix: Add this.sessionClusterKey (which defaults to the application name) that will be used for the sessionID and when getting the instance of the session "Cache". This allows for different instances of the same application with different application names to use the same session cache provided they each have this.sessionClusterKey="VALUE" to the same value.