When using <cfset el = XMLElemNew(doc,"PREFIX:ELEMENT") /> the PREFIX does not register and the namespace is not applied.
This fix will get any xmlns prefix declarations from the document root and apply the namespace to the new element with the prefix. Works normally if there is no prefix specified or the prefix does not match a namespace declaration.

ACF does this, but it actually appears to get xmlns prefixes from anywhere in the document path up to the root. in other words, thich change brings railo closer to ACF compatibility, but not quite.