Update Jira1441.cfc
It appears the error only occurs when: you group for more than two levels (i.e. 2 or more nested cfoutputs) and when your maxrows is set to a number greater than 1.

I'm not able to run the test cases so did my best to write it and test it in a normal file and then make modifications to this one. However you may need to make minor modifications to this file.